feat: implement route protection and refactor components into individual files
